Update reaction controller

+++ src/main/java/CS544/Controller/ReactionController.java
@@ -1,6 +1,5 @@
 package CS544.Controller;
 
-import CS544.Dao.IReactionDao;
 import CS544.Model.Post;
 import CS544.Model.Reaction;
 import CS544.Model.User;
@@ -8,18 +7,13 @@ import CS544.Service.PostService;
 import CS544.Service.ReactionService;
 import CS544.Service.UserService;
 import io.jsonwebtoken.Claims;
-import jakarta.annotation.Resource;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.view.RedirectView;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping( "/reaction")
@@ -70,15 +64,25 @@ public class ReactionController {
     }
 
     @PutMapping(value = "/update/{id}")
-    public void put(@PathVariable long id, @Valid @RequestBody Reaction reaction){
-        if (id != reaction.getId()) {
-            throw new IllegalArgumentException();
+    public ResponseEntity<Reaction> put(@PathVariable long id, @Valid @RequestBody Reaction reaction){
+        Reaction reaction1 = reactionService.getReaction(id);
+        if(reaction1==null){
+            return ResponseEntity.notFound().build();
         }
-        reactionService.updateReaction(reaction);
+        reaction1.setReacted(reaction.getReacted());
+        reactionService.updateReaction(reaction1);
+        return ResponseEntity.ok(reaction1);
     }
 
     @DeleteMapping(value = "/delete/{id}")
-    public void delete(@PathVariable long id){
+
+    public ResponseEntity<String> delete(@PathVariable long id){
+        Reaction reaction1 = reactionService.getReaction(id);
+        if(reaction1==null){
+            return ResponseEntity.notFound().build();
+        }
         reactionService.deleteReaction(id);
+        return ResponseEntity.ok("Reaction updated successfully!!");
     }
+
 }
